Support character maps during serialization
This one is messy. There are several bugs involved. First, XML Calabash was not
attempting to preserve the character maps property from the XSLT step. That’s
fixed. There’s a dichotomy in the specs: XSLT uses QNames to refer to character
maps, but the fn:serialize function uses a direct map. XML Calabash stores the
character map in document properties using the latter format. You can, as a
consequence, construct that directly for use in XML Calabash serialization if
you like.

There are also bugs in Saxon 12; see
[#6860](https://saxonica.plan.io/issues/6860).
I’ve rewritten the serializer to work around those bugs.