Fix errors in visualizer error output; close .dot files
Fix #120

A couple of things here.

1. The visualizer doesn’t have access to a step configuration, so it was
outputting messages to the *underlying* logger. That’s only accessible if you actually
fiddle with the logback configuration. I’ve reworked it so that the error
messages are reported by the message reporter so it’s easier to see what they are.
2. The error on Windows was that the .dot file was inaccessible because I’d
failed to close it after writing to it.