Correct the documentation for output filespecs; fix appending
Fix #123

The documentation incorrectly described how to create a filespec number of a
specific width. It uses String.format style padding, not format-number style
padding. I also fixed the implementation so that it appends multiple results to
the same file, as described in the documentation, instead of replacing the file each time.