Make sure the base URI of non-XML documents are computed correctly
Fix #138
Fix #139

Data models constructed for non-XML inputs were accidentally using the base URI
of the static context instead of their actual base URI. In the case of #138,
this meant that the fn:base-uri() function returned the wrong value. In the case
of #139, it meant that the content type of the document was incorrect.