Make changing the logback configuration file easier
This PR adds a logback-debug.xml file to the lib directory in the distribution.
This file prints debug messages to the console and saves them in a file named
xmlcalabash.log. It also updates the included scripts so that arguments of the
form -D… are recognized as Java property settings and added to the command line
in the right place. (Note that on Windows, you have to quote any property
setting that includes a “.”, for example
'-Dlogback.configurationFile=lib/logback-debug.xml')