First attempt at a BaseX implementation for XQuery processing
Fix #190

This PR adds a cx:basex step; it also adds configuration options to allow a user
to specify that BaseX should be the default processor for the p:xquery step.
This supports both local databases and connecting to a remote database.

It’s a fairly large PR in part because it adds all of the infrastructure
necessary to *have* separate implementations of XQuery. Future additions, such
as an eXist-db implementation should be smaller and simpler.

There’s some documentation, but perhaps not enough. Issues raised to that effect
most welcome.

Hat tip to Christian Grün for patiently helping me figure out some of the BaseX
API details. An errors that remain are wholly my own.

Hat tip also to Amanda Galtman for making the suggestion. It’s obvious in
retrospect but it had never occurred to me.