Attempt to fix (Windows) console encoding problems
Fix #199

As far as I can tell, the root cause here is that in Java 18, the default
character set was changed to UTF-8, irrespective of the platform. So if you run
XML Calabash with Java 18+ on Windows, Java says the default character set is
UTF-8 and it often just isn’t.

The current workaround is to use -Dfile.encoding=COMPAT if no explicit
file.encoding property is specified. That gives backwards-compatible behavior.
Alternate solutions: explaining how to set the Windows encoding to UTF-8 or
adding an XML Calabash configuration property to specify the default encoding
might be better in the long term.

If you *have* changed your shell to UTF-8 on Windows (or some other platform)
and file.encoding=COMPAT does the wrong thing, please let me know. *Your*
workaround is to pass the file encoding to the processor explicitly.

I also added a line to the “info version” output so that it’s easy to tell what
the JVM thinks the default character set is.

And in an only tangentially related area, I changed the Java logger so that it
doesn’t log to the console. Logging error messages to the console created double
messages because XML Calabash also logs error messages to the console. The Java
logger probably makes more sense if you want to log to a file or if you’re
running in some application server environment where logging is centralized.
Anyway, see https://docs.xmlcalabash.com/userguide/current/logging.html for more
detail.