Always attempt to resolve static option values statically
Fix #305

The NPE reported arises because the expression did not appear to be resolvable
statically, so it wasn’t. But for a static option initializer, we should always
try. If the expression can’t be resolved statically, an error will occur.

The particular issue in this expression was that it contained a call to
p:system-property which is not considered statically resolvable because the
'p:episode' argument isn’t. I’ve improved the analysis so that for simple,
literal strings (which I expect is the 99% case), other values are considered
static.