Rework the Maven dependencies for the app and xmlcalabash packages
Fix #304

This PR removes the dependency on Saxon-EE which was never intended.

It also removes many of the dependencies needed for optional steps.
For example, all of the Selenium dependencies are no longer
transitively required. That will mean that you’ll need to add them to
your build system if you’re using those steps, but it makes the Maven
distributions much cleaner.

All of the dependencies are still bundled with the released
application, these changes only apply to projects that are using Maven
dependencies to include XML Calabash.