Synchronize access to the ExpressionStep.run method
Fix #340

It’s clear that the expression step isn’t thread safe because the underlying
expressions can be shared across different steps. In this case, the message
select expression is the same expression in both instances of the
mySteps:parallelism step (because there’s only one instance of it in the
pipeline).

Is this unique to ExpressionStep? It could be, but I’m not sure. It’s worth
investigating if separate instances of the underlying expressions can be
duplicated.