Keep track of the actual static options encountered
Fix #380

We compute all the options provided on the command line before we begin
compiling. Any static option that we encounter during compilation will use the
appropriate static value, but we don’t then want to try to set that option at
runtime. But we were failing to keep track of the static options encountered.