Actually, at last, support the dtd-attribute-values
This PR does two things. First, it actually adds support for the
dtd-attribute-values option on p:validate-with-relax-ng. Hat tip to George Bina
for doing the actual implementation work in Jing and for getting me out of the
weeds.

I’ve also added support for recognizing that XVRL parameters were passed to the
various flavors of validation step. They’re still ignored, but at least there’s
a framework for recognizing them.