Construct SAX sources directly from XdmNodes, preserving locations
Fix #419

This PR includes a new utility class that walks an XdmNode tree directly
producing SAX events. This allows it to preserve the location information in the
initial document. It can be used anywhere that a SAXSource is allowed, but not,
alas, anywhere an InputSource is used. Sadly, the Jing schema parser accepts only
an InputSource which it immediately wraps in a SAXSource.