Correct a case where deadlock could occur; fail if deadlocked
Fix #414

This PR fixes a case where all of the steps might be unready to run. It also
adds a check for that case and fails with an error.

The bug was caused by a confluence of events: a user defined step with an input
that could be empty and calls to that step within a loop where it’s explicitly
bound to empty.