Make sure empty ports are tracked per-instance, not per type
Fix #445

Supporting default inputs on ports and the constraint that an explicit p:empty
is not the same as “not provided” is tricky. If a port is bound to p:empty, it’s
marked as “welded shut” in the data model so that we don’t wait for input to
arrive on it. But in the case of user-defined steps, that was being tracked on
the implementation of the step, not the instance of the step. But all instances
of a user-defined step share the same implementation so mixing p:empty with
other bindings could cause deadlock.

I think a better implementation might have been to have an EmptyStep in the
graph that simply closed its result port without sending any documents on it.
The extra bookkeeping about welded ports was an effort to avoid that extra step.
I’m not sure it was a good trade off.