Fix the formatting of RELAX NG tableaux
For reasons I do not understand, Firefox version 146.0.1 on macOS is
ignoring or discarding <br> elements inside code. At least that’s what
appears to be happening. It works fine on Linux, so there’s something
odd happening. I don’t actually need the code wrapper, so I’ve
replaced it with a span and updated the CSS. That seems to fix the
problem, but honestly ¯\_(ツ)_/¯